requirements.txt
gtts==2.3.2
Pillow==10.0.0
playsound==1.2.2
